session storage gets reset during authorization
